Refactoring to condense.
